Finalized migrating validation
